Check in many in trx
